Former-commit-id: 893e75715219c0f6df08ac8392dbd72f2bf606d1

+++ Intersect Editor/Forms/Editors/frmClass.cs
@@ -878,7 +878,7 @@ namespace Intersect.Editor.Forms
 
         private void nudExpIncrease_ValueChanged(object sender, EventArgs e)
         {
-            _editorItem.ExpIncrease = (long) nudExpIncrease.Value;
+            _editorItem.ExpIncrease = (int) nudExpIncrease.Value;
         }
 
         private void nudHpIncrease_ValueChanged(object sender, EventArgs e)
@@ -934,7 +934,7 @@ namespace Intersect.Editor.Forms
 
         private void nudBaseExp_ValueChanged(object sender, EventArgs e)
         {
-            _editorItem.BaseExp = (long) nudBaseExp.Value;
+            _editorItem.BaseExp = (int) nudBaseExp.Value;
         }
 
         private void nudBaseHP_ValueChanged(object sender, EventArgs e)
